backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai-whisper==20240930
numpy==1.26.0
pydub==0.25.1
requests==2.31.0
python-dotenv==1.0.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0